+++ src/grafana/provisioning/datasources/datasources.yaml
@@ -1,7 +1,7 @@
 apiVersion: 1
 datasources:
   - name: ClickHouse
-    type: grafana-clickhouse-datasource
+    type: ClickHouse
     jsonData:
       defaultDatabase: default
       port: 9000
